type everything and restructuralize utils
