+++ src/lib/utils/constants.js
@@ -1,0 +1,3 @@
+export const MODE_DECADE = 0;
+export const MODE_YEAR = 1;
+export const MODE_MONTH = 2;
