--- src/lib/utils/settings.js
@@ -1,19 +1,0 @@
-import { en } from '../i18n/index.js';
-
-export default {
-  theme: 'sdt-calendar-colors',
-  mode: 'auto',
-  format: 'yyyy-mm-dd',
-  formatType: 'standard',
-  displayFormat: null,
-  displayFormatType: null,
-  minuteIncrement: 1,
-  weekStart: 1,
-  inputClasses: '',
-  todayBtnClasses: 'sdt-action-btn sdt-today-btn',
-  clearBtnClasses: 'sdt-action-btn sdt-clear-btn',
-  todayBtn: true,
-  clearBtn: true,
-  autoclose: true,
-  i18n: en
-}
